streaming.csv
revised version of movies_on_netflix_prime_video_hulu_and_disney.csv.  The newest csv for the streaming data is called, "Streaming" in the schema and in PostgreSQL so I needed to revise the original csv to mach the schema that we loaded into PostgreSQL. One thing to note, however, is the the schema lists Year as date format.  That doesn't work with the csv file because the original source data never provided month and day, and without that the year can only be conveyed as a int or VARCHAR.